LPS-182767 Extract FDSDataProvider methods in CommerceCategoryDisplayPageTableFDSView to a new CommerceCategoryDisplayPageFDSDataProvider component
